requirements.txt
fastapi
uvicorn
jinja2
textual
python-vlc
yt-dlp
sqlalchemy
rich
python-multipart